feat: Add TypingText component for animated text display
This commit adds a new component called TypingText to the note directory. The TypingText component uses the gsap library and ScrollTrigger plugin to create an animation effect where two text elements, "王安琪" and "胡奕", move horizontally and rotate in a loop. The animation is triggered when the component mounts. This new component enhances the visual appeal and interactivity of the note page.